docs/_code/building-secure-mcp-server-for-weather/pyproject.toml
[project]
name = "building-secure-mcp-server-for-weather"
version = "0.1.0"
description = "MCP Server For Weather"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "uvicorn[standard]",
    "fastapi",
    "requests"
]